Addressing review comments, rebase
