+++ rust/lance-index/src/lib.rs
@@ -21,6 +21,8 @@ use lance_core::Result;
 pub mod scalar;
 pub mod vector;
 
+pub const INDEX_FILE_NAME: &str = "index.idx";
+
 pub mod pb {
     #![allow(clippy::use_self)]
     include!(concat!(env!("OUT_DIR"), "/lance.index.pb.rs"));
@@ -34,4 +36,24 @@ pub trait Index: Send + Sync {
     fn as_index(self: Arc<Self>) -> Arc<dyn Index>;
     /// Retrieve index statistics as a JSON string
     fn statistics(&self) -> Result<String>;
+    /// Get the type of the index
+    fn index_type(&self) -> IndexType;
+}
+
+/// Index Type
+pub enum IndexType {
+    // Preserve 0-100 for simple indices.
+    Scalar = 0,
+    // 100+ and up for vector index.
+    /// Flat vector index.
+    Vector = 100,
+}
+
+impl std::fmt::Display for IndexType {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            Self::Scalar => write!(f, "Scalar"),
+            Self::Vector => write!(f, "Vector"),
+        }
+    }
 }
