+++ rust/lance/benches/ivf_pq.rs
@@ -20,10 +20,11 @@ use criterion::{criterion_group, criterion_main, Criterion};
 use lance::{
     arrow::*,
     dataset::{WriteMode, WriteParams},
-    index::{vector::VectorIndexParams, DatasetIndexExt, IndexType},
+    index::{vector::VectorIndexParams, DatasetIndexExt},
     Dataset,
 };
 
+use lance_index::IndexType;
 use lance_linalg::distance::MetricType;
 use lance_testing::datagen::generate_random_array;
 #[cfg(target_os = "linux")]
