+++ rust/lance/benches/vector_index.rs
@@ -23,6 +23,7 @@ use arrow_array::{
 use arrow_schema::{DataType, Field, FieldRef, Schema as ArrowSchema};
 use criterion::{criterion_group, criterion_main, Criterion};
 use futures::TryStreamExt;
+use lance_index::IndexType;
 #[cfg(target_os = "linux")]
 use pprof::criterion::{Output, PProfProfiler};
 use rand::{self, Rng};
@@ -31,7 +32,7 @@ use lance::dataset::{WriteMode, WriteParams};
 use lance::index::vector::ivf::IvfBuildParams;
 use lance::index::vector::pq::PQBuildParams;
 use lance::index::vector::VectorIndexParams;
-use lance::index::{DatasetIndexExt, IndexType};
+use lance::index::DatasetIndexExt;
 use lance::{arrow::as_fixed_size_list_array, dataset::Dataset};
 use lance_arrow::FixedSizeListArrayExt;
 use lance_linalg::distance::MetricType;
